*** empty log message ***

+++ src/java/se/idega/idegaweb/commune/school/report/business/NackaCCPreSchoolPlacementReportModel.java
@@ -1,5 +1,5 @@
 /*
- * $Id: NackaCCPreSchoolPlacementReportModel.java,v 1.14 2004/01/27 12:07:38 anders Exp $
+ * $Id: NackaCCPreSchoolPlacementReportModel.java,v 1.15 2004/01/27 14:15:29 anders Exp $
  *
  * Copyright (C) 2003 Agura IT. All Rights Reserved.
  *
@@ -22,10 +22,10 @@ import com.idega.block.school.data.SchoolArea;
 /** 
  * Report model for child care pre school placements in Nacka.
  * <p>
- * Last modified: $Date: 2004/01/27 12:07:38 $ by $Author: anders $
+ * Last modified: $Date: 2004/01/27 14:15:29 $ by $Author: anders $
  *
  * @author Anders Lindman
- * @version $Revision: 1.14 $
+ * @version $Revision: 1.15 $
  */
 public class NackaCCPreSchoolPlacementReportModel extends ReportModel {
 
@@ -477,7 +477,7 @@ public class NackaCCPreSchoolPlacementReportModel extends ReportModel {
 	/**
 	 * Returns the number of child placements for the specified school.
 	 */
-	protected int getProviderMeanHours(int schoolId) throws RemoteException {
+	protected float getProviderMeanHours(int schoolId) throws RemoteException {
 		ReportBusiness rb = getReportBusiness();
 		int schoolType1 = rb.getPreSchoolTypeId();
 		int schoolType2 = rb.getGeneralPreSchoolTypeId();
